+++ CHANGELOG.md
@@ -9,3 +9,11 @@
 ## 0.0.3
 
 * Bug fixes and improvements
+
+## 0.0.4
+
+* Update classes names
+
+## 0.0.5
+
+* Add onCallEngineSuccessfully callback function